feat: add Shizuku input injection, cursor overlay, and protocol fixes
- Add Shizuku-based input injection for non-root mouse/keyboard control
- Implement floating cursor overlay with SYSTEM_ALERT_WINDOW permission
- Fix key event parsing: keysym is in first 2 bytes, not last
- Add keysym normalization for InputLeap's 0xEFxx -> 0xFFxx range
- Fix cursor overlay threading (UI ops must run on main thread)
- Add cursor overlay toggle in Settings with permission request flow
- Pre-start CursorOverlayService to avoid instance timing issues
- Add FLAG_LAYOUT_IN_SCREEN for correct overlay positioning

+++ app/src/main/java/com/inputleaf/android/uhid/UhidEventSocket.kt
@@ -44,12 +44,12 @@ class UhidEventSocket {
                     is InputLeapEvent.KeyDown -> {
                         KeysymTable.toHidUsage(event.keyId) ?: return
                         o.writeByte(0x01); o.writeInt(event.keyId)
-                        o.writeByte(0x00); o.writeByte(event.mask.toByte())
+                        o.writeByte(0x00); o.writeByte(event.mask.toInt())
                     }
                     is InputLeapEvent.KeyUp -> {
                         KeysymTable.toHidUsage(event.keyId) ?: return
                         o.writeByte(0x01); o.writeInt(event.keyId)
-                        o.writeByte(0x01); o.writeByte(event.mask.toByte())
+                        o.writeByte(0x01); o.writeByte(event.mask.toInt())
                     }
                     is InputLeapEvent.MouseMoveRel -> {
                         o.writeByte(0x02); o.writeInt(event.dx); o.writeInt(event.dy)
